requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
azure-core==1.30.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.1
googletrans==3.0.0
grpcio==1.64.1
grpcio-status==1.62.2
gspread==6.1.2
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
icecream==2.1.3
idna==2.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
msrest==0.7.1
multidict==6.0.5
numpy==2.0.0
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyimgur==0.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
